[Refactor] lightning and scripts folder
