--- datamodule/README.md
@@ -1,1 +1,0 @@
-These are the LightningDataModule implmented for the various datasets.
