--- dataset/README.md
@@ -1,1 +1,0 @@
-The PyTorch Dataset Implementations.
