--- model/README.md
@@ -1,1 +1,0 @@
-Various Neural Network Models being one model per folder.
